Move message events to own file
