+++ src/entities/message.rs
@@ -45,99 +45,6 @@ pub struct Message {
     role_subscription_data: Option<RoleSubscriptionData>,
 }
 
-#[derive(Debug, Serialize, Deserialize, Default)]
-pub struct MessageCreate {
-    #[serde(flatten)]
-    message: Message,
-    guild_id: Option<Snowflake>,
-    member: Option<GuildMember>,
-    mentions: Vec<(User, GuildMember)>, // Not sure if this is correct: https://discord.com/developers/docs/topics/gateway-events#message-create
-}
-
-impl WebSocketEvent for MessageCreate {}
-
-#[derive(Debug, Serialize, Deserialize, Default)]
-pub struct MessageUpdate {
-    #[serde(flatten)]
-    message: Message,
-    guild_id: Option<String>,
-    member: Option<GuildMember>,
-    mentions: Vec<(User, GuildMember)>, // Not sure if this is correct: https://discord.com/developers/docs/topics/gateway-events#message-create
-}
-
-impl WebSocketEvent for MessageUpdate {}
-
-#[derive(Debug, Serialize, Deserialize, Default)]
-pub struct MessageDelete {
-    id: String,
-    channel_id: String,
-    guild_id: Option<String>,
-}
-
-impl WebSocketEvent for MessageDelete {}
-
-#[derive(Debug, Serialize, Deserialize, Default)]
-pub struct MessageDeleteBulk {
-    ids: Vec<String>,
-    channel_id: String,
-    guild_id: Option<String>,
-}
-
-impl WebSocketEvent for MessageDeleteBulk {}
-
-#[derive(Debug, Serialize, Deserialize, Default)]
-pub struct MessageReactionAdd {
-    user_id: String,
-    channel_id: String,
-    message_id: String,
-    guild_id: Option<String>,
-    member: Option<GuildMember>,
-    emoji: Emoji,
-}
-
-impl WebSocketEvent for MessageReactionAdd {}
-
-#[derive(Debug, Serialize, Deserialize, Default)]
-pub struct MessageReactionRemove {
-    user_id: String,
-    channel_id: String,
-    message_id: String,
-    guild_id: Option<String>,
-    emoji: Emoji,
-}
-
-impl WebSocketEvent for MessageReactionRemove {}
-
-#[derive(Debug, Serialize, Deserialize, Default)]
-pub struct MessageReactionRemoveAll {
-    channel_id: String,
-    message_id: String,
-    guild_id: Option<String>,
-}
-
-impl WebSocketEvent for MessageReactionRemoveAll {}
-
-#[derive(Debug, Serialize, Deserialize, Default)]
-pub struct MessageReactionRemoveEmoji {
-    channel_id: String,
-    message_id: String,
-    guild_id: Option<String>,
-    emoji: Emoji,
-}
-
-impl WebSocketEvent for MessageReactionRemoveEmoji {}
-
-#[derive(Debug, Deserialize, Serialize, Default)]
-pub struct TypingStartEvent {
-    pub channel_id: String,
-    pub guild_id: Option<String>,
-    pub user_id: String,
-    pub timestamp: i64,
-    pub member: Option<GuildMember>,
-}
-
-impl WebSocketEvent for TypingStartEvent {}
-
 #[derive(Debug, Serialize, Deserialize)]
 pub struct MessageReference {
     pub message_id: Snowflake,
